requirements.txt
langchain-astradb
langchain 
datasets
pypdf
python-dotenv
chainlit
langchain-google-genai